requirements.txt
numpy
gymnasium
tensorboard
matplotlib
pillow
